+++ srcs/start_printf.c
@@ -50,7 +50,7 @@ void	check_type(const char *format, t_args *ar, va_list *ap)
 		|| format[ar->index] == 'D')
 	{
 	 	ar->conv = format[ar->index];
-		ar->num = va_arg(*ap, long long);
+		ar->num = va_arg(*ap, long);
 		check_int(ar);
 	}
 	else if (format[ar->index] == 'o' || format[ar->index] == 'O')
